requirements.txt
python-telegram-bot==20.7
google-generativeai
supabase
python-dotenv
dateparser
langchain
langchain-google-genai
langchain-core